[kernel][crashlogs] Handle spontaneous reboot
Start to make use of the new ram_crashlog in order to maintain a
partial crashlog (platform permitting) at all times, so that if we
suffer a spontaneous reboot for any reason.

In addition to enabling this functionality, there is a lot of general
cleanup and consolidation in this CL.  Some duplicated code has been
centralized, some dead code has been removed, the HW watchdog code has
been renamed in order to distinguish it from the software watchdog
code (which is technically dead/inactive right now), and so on.

There is still a bunch of cleanup which can be done, but that can come
later.

Bug: 38142
Tested: manually forced crashes on Sherlock using the kernel shell.
Specifically
++ k crash
++ k crash with kernel.halt-on-panic set
++ k pmm oom
++ k wdt force

Also unit tests for the boot log checker (fx run-test boot_log_checker)

Change-Id: I310ff0f22371414854f42df8adb1a6a83acfbcc5