[zbitl] Support EFI File Protocol
This provides a zbitl::StorageTraits specialization for using
an EFI File Protocol handle as ZBI storage.  It also adds some
mock/fake support for the protocol in the efi/testing library
that's used to enable the full zbitl tests on the new backend.

As zbitl is already compiled in the kernel.phys32 environment,
this needed to clean up the EFI headers slightly to be fully
compatible with 32-bit environments (which should actually
work for 32-bit UEFI if that's ever needed).

Bug: 98798
Change-Id: I157db9506d4f488731ef07e133b252282fa0f201
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677302
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>